fix: SpaceCart function moved to SpaceCart module to allow for re-renering of Html based upon future button click event listener
